Use a jdk, use container support
This shouldn't take 100s of mb of memory, so add the container support
to limit the heap based on the amount of memory allocated to the
container. Tested with 48m heap.
Use a jdk instead of just a jre in case we need to do additional
troubleshooting in a container